pyproject.toml
[build-system]
requires = ["setuptools>=80.10"]
build-backend = "setuptools.build_meta"

[project]
name = "django-stratagem"
version = "2026.2.1b0"
description = "Registry-based plugin architecture for Django applications"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Jack Linke", email = "jack@watervize.com" },
]
keywords = ["django", "registry", "plugin", "strategy", "pattern", "extensible"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
    "Framework :: Django :: 6.0",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries",
    "Topic :: Utilities",
]
dependencies = [
    "django>=4.2",
]

[project.optional-dependencies]
drf = [
    "djangorestframework>=3.14",
]
waffle = [
    "django-waffle>=4.0",
]

[project.urls]
Homepage = "https://github.com/OmenApps/django-stratagem"
Repository = "https://github.com/OmenApps/django-stratagem"
Issues = "https://github.com/OmenApps/django-stratagem/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
src = ["src"]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["django_stratagem"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
django_find_project = false
pythonpath = ["."]
python_files = ["test_*.py"]
log_cli = true

[tool.coverage.run]
branch = true
source_pkgs = ["django_stratagem"]

[tool.coverage.report]
show_missing = true
omit = [
    ".nox/*",
    "*/migrations/*",
    "*/__init__.py",
]

[tool.coverage.paths]
source = [
    "src",
    "*/site-packages",
]

[dependency-groups]
dev = [
    "bandit",
    "coverage[toml]",
    "djangorestframework",
    "nox",
    "pip-audit",
    "pre-commit",
    "pre-commit-hooks",
    "pytest",
    "pytest-cov",
    "pytest-django",
    "pytest-mock",
    "pytest-subtests",
    "pytest-timeout",
    "pyupgrade",
    "ruff",
]
docs = [
    "furo==2024.8.6",
    "linkify-it-py==2.0.3",
    "myst-parser==4.0.1",
    "sphinx==8.2.3",
    "sphinx-copybutton==0.5.2",
    "sphinx-togglebutton==0.3.2",
    "sphinxcontrib-mermaid==1.0.0",
]
